backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0
google-generativeai==0.3.0
scikit-learn==1.0.2
numpy==1.21.2
python-multipart==0.0.5
requests==2.26.0
typing-extensions==3.10.0.2
youtube-transcript-api==0.6.2 